# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rock-n-wall api",
        "description": "Documentation Rock-n-wall API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://5.35.83.190",
            "description": "Debug test server"
        }
    ],
    "paths": {
...
        "/api/v1/essences/{essenceId}/sub-essences/{subEssenceId}": {
            "get": {
                "tags": [
                    "Pages"
                ],
                "summary": "Получить детальные данные для страницы sub essence",
                "operationId": "dd2491bed5e0d4e554e6119c719fd170",
                "parameters": [
                    {
                        "name": "essenceId",
                        "in": "path",
                        "description": "Essence id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "subEssenceId",
                        "in": "path",
                        "description": "Sub essence id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные для детальной страницы sub essence",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubEssenceShow"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth_api": []
                    }
                ]
            }
        },
        "/api/v1/essences/{essenceId}/sub-essences/{subEssenceId}/paginate": {
            "get": {
                "tags": [
                    "Paginate"
                ],
                "summary": "Пагинация изображений на детальной странице subEssence",
                "operationId": "8d446cf4127dd2b6e3e9537ed9acb53f",
                "parameters": [
                    {
                        "name": "essenceId",
                        "in": "path",
                        "description": "Essence id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "subEssenceId",
                        "in": "path",
                        "description": "Sub essence id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Последний или первый id изображения в текущем sub essence (первый если надо предыдущие получить, последний если следующие)",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "need",
                        "in": "query",
                        "description": "Необходимо получить следующие или предыдущие",
                        "schema": {
                            "type": "integer"
                        },
                        "example": "next/prev"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Следующие или предыдущие изображения на странице sub essence",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth_api": []
                    }
                ]
            }
        },
        "/api/v1/sub-essences/{subEssenceId}/paginate": {
            "get": {
                "tags": [
                    "Paginate"
                ],
                "summary": "Короткий запрос для пагинации изображений на детальной странице subEssence",
                "operationId": "e6c9a91ba53876b595c0c0ebbe7880e9",
                "parameters": [
                    {
                        "name": "subEssenceId",
                        "in": "path",
                        "description": "Sub essence id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Последний или первый id изображения в текущем sub essence (первый если надо предыдущие получить, последний если следующие)",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "need",
                        "in": "query",
                        "description": "Необходимо получить следующие или предыдущие",
                        "schema": {
                            "type": "integer"
                        },
                        "example": "next/prev"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Следующие или предыдущие изображения на странице sub essence",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth_api": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "description": "Product data",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "string",
                        "example": "123"
                    },
                    "name": {
                        "type": "string",
                        "format": "string",
                        "example": "Black car"
                    },
                    "vip": {
                        "type": "boolean",
                        "format": "bool",
                        "example": "true"
                    },
                    "live": {
                        "type": "boolean",
                        "format": "bool",
                        "example": "true"
                    },
                    "image": {
                        "type": "string",
                        "format": "string",
                        "example": "http://domain/storage/images/black-car.jpg"
                    }
                },
                "type": "object"
            },
...
            "SubEssenceShow": {
                "description": "Sub essence show data",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "string",
                        "example": "123"
                    },
                    "name": {
                        "type": "string",
                        "format": "string",
                        "example": "Cats"
                    },
                    "image": {
                        "type": "string",
                        "format": "string",
                        "example": "http://domain/storage/images/cats.jpg"
                    },
                    "displayType": {
                        "type": "string",
                        "format": "string",
                        "example": "selections/categories"
                    },
                    "products": {
                        "type": "array",
                        "format": "array",
                        "items": {
                            "$ref": "#/components/schemas/Product"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "auth_api": {
                "type": "apiKey",
                "name": "app-key",
                "in": "header"
            }
        }
    },
...
}